- fix some deserialization issues - some tidy up - code check in for raya 0.2

+++ pesta/pesta/Utilities/UnixTime.cs
@@ -29,11 +29,11 @@ namespace Pesta.Utilities
             return origin.AddSeconds(timestamp);
         }
 
-        static public double ConvertToUnixTimestamp(DateTime date)
+        static public long ConvertToUnixTimestamp(DateTime date)
         {
             DateTime origin = new DateTime(1970, 1, 1, 0, 0, 0, 0);
             TimeSpan diff = date - origin;
-            return Math.Floor(diff.TotalSeconds);
+            return (long)Math.Floor(diff.TotalSeconds);
         }
     }
 }
